Move IsSimpleOp into ./valid next to IsValidOp (match JTS)
